+++ packages/ui-components/web/src/Components/Button.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import type { CSSProperties } from 'react';
 import classNames from 'classnames';
 import type { RichContentTheme } from 'wix-rich-content-common';
 import { mergeStyles } from 'wix-rich-content-common';
@@ -18,6 +19,7 @@ export interface ButtonProps {
   secondary?: boolean;
   borderless?: boolean;
   theme: RichContentTheme;
+  style?: CSSProperties;
 }
 
 const Button: React.FC<ButtonProps> = ({
@@ -32,6 +34,7 @@ const Button: React.FC<ButtonProps> = ({
   borderless = false,
   theme,
   isMobile = false,
+  style,
 }) => {
   const styles = mergeStyles({ styles: Styles, theme });
   const buttonStyle = secondary ? styles.button_secondary : styles.button_primary;
@@ -46,6 +49,7 @@ const Button: React.FC<ButtonProps> = ({
         [styles.borderLess]: borderless,
         [styles.mobile]: isMobile,
       })}
+      style={style}
     >
       {children}
     </button>